Fixed warnings, non-compiling code and typos
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2791 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/server/CServer_Parse.cpp
@@ -1287,7 +1287,7 @@ void GameServer::ParseConnect(NetworkSocket tSocket, CBytestream *bs) {
 				avgLives = avgLives / wormCount * tLXOptions->tGameinfo.iAllowConnectDuringGameLivesPercent / 100.0f;
 				if( avgLives < 0 || tLXOptions->tGameinfo.iAllowConnectDuringGameLivesPercent <= 0 )
 					avgLives = WRM_OUT;
-				newcl->getWorm(i)->setLives( avgLives );
+				newcl->getWorm(i)->setLives( (int)avgLives );
 			};
 			newcl->getWorm(i)->setKills(0);
 			newcl->getWorm(i)->setGameScript(cGameScript.get());
